requirements.txt
alembic==1.11.3
aniso8601==9.0.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
asyncpg==0.28.0
attrs==23.1.0
autopep8==2.0.2
bcrypt==3.2.0
blinker==1.6.2
build==0.10.0
CacheControl==0.13.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cleo==2.0.1
click==8.1.3
colorama==0.4.6
config==0.5.1
crashtest==0.4.1
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.7
Django==4.2.11
django-crispy-forms==2.0
django-formtools==2.4.1
django-import-export==3.2.0
django-otp==1.2.2
django-phonenumber-field==6.4.0
django-two-factor-auth==1.15.4
django-widget-tweaks==1.4.12
docopt==0.6.2
dulwich==0.21.5
et-xmlfile==1.1.0
filelock==3.12.2
flake8==6.1.0
greenlet==2.0.2
idna==3.4
importlib-metadata==6.8.0
install==1.3.5
installer==0.7.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
Jinja2==3.1.2
jsonschema==4.17.3
kaleido==0.2.1
keyring==24.2.0
Mako==1.2.4
MarkupPy==1.14
MarkupSafe==2.1.3
mccabe==0.7.0
more-itertools==10.1.0
msgpack==1.0.5
numpy==1.25.1
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
pandas==1.3.4
pexpect==4.8.0
phonenumbers==8.13.18
pipdeptree==2.12.0
pipreqs==0.4.13
pkginfo==1.9.6
platformdirs==3.9.1
plotly==5.16.1
poetry-core==1.7.0
psycopg2==2.9.1
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
pypng==0.20220715.0
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==0.19.1
pytz==2023.3
pywin32-ctypes==0.2.2
PyYAML==6.0.1
qrcode==7.4.2
rapidfuzz==2.15.1
requests==2.31.0
requests-toolbelt==1.0.0
shellingham==1.5.3
six==1.16.0
SQLAlchemy==1.4.22
sqlparse==0.4.4
tablib==3.5.0
tabulate==0.9.0
tenacity==8.2.3
tomli==2.0.1
tomlkit==0.12.1
trove-classifiers==2023.8.7
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.4
vanna==0.0.21
virtualenv==20.24.2
Werkzeug==2.3.4
WTForms==3.0.1
xlrd==2.0.1
xlwt==1.3.0
yapf==0.40.1
yarg==0.1.9
zipp==3.16.2
